chore: move functions from utils to more specific files
